aqs-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
SQLAlchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
python-dotenv==1.1.0
gunicorn==22.0.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.0.1
python-jose[cryptography]==3.4.0
email-validator==2.1.0.post1
psycopg2-binary==2.9.9